WiP: Add referral bonus when user use "top-up wallet" + minimum top-up amount needed for bonus
Still WiP, but I might need your help with furter instructions. This new functionality has been implemented in the referral program that allows the referral bonus to be awarded when referred user top-up's their wallet. Correct me if I'm wrong, but as far as I know this is currently only possible when referred user buys Moodle course, right? Also I added a setting for the minimum amount of the required top-up in order to obtain a reference credit has been added.