compability with paygw_bank plugin
The wallet plugin was throwing an exception when trying to retrieve payment information for (bank9 transactions that were no longer present in the enrol_wallet_items table. This table is used for temporary storage of payment information, but the records are deleted after a certain period. When the bank payment gateway tried to access this information later (e.g., for approval or management purposes), it would fail if the record had been deleted